LayoutTests/http/tests/security/xssAuditor/xss-protection-parsing-04-expected.txt

CONSOLE MESSAGE: Refused to execute a JavaScript script. Source code of script found within request.

This tests that the X-XSS-Protection header is not ignored when the report and mode directives are swapped.



--------
Frame: 'frame'
--------
